pyproject.toml
[tool.poetry]
name = "orc-reader"
version = "0.1.0"
description = ""
authors = ["nguyen van duong <nguyen.van.duong@sun-asterisk.com>"]

[tool.poetry.dependencies]
python = "^3.6"
click = "^7.0"
yapf = "^0.30.0"
isort = "^5.7.0"
Pillow = "^8.1.0"
pytesseract = "^0.3.7"
numpy = "^1.19.5"
opencv-python = "^4.5.1"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
